requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
docutils==0.18.1
fonttools==4.31.1
idna==3.3
Kivy==2.1.0
Kivy-Garden==0.1.4
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pkg_resources==0.0.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
scipy==1.8.0
six==1.16.0
urllib3==1.26.9
